BlobTriggerFunction/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
python-gnupg
pgpy
